LayoutTests/http/tests/security/sandboxed-iframe-blocks-access-from-parent-expected.txt

CONSOLE MESSAGE: line 12: SecurityError: Sandbox access violation: Blocked a frame at "http://127.0.0.1:8000" from accessing a cross-origin frame.  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
Sandboxing a frame puts it into a unique origin by default, which the containing document shouldn't have script access to. This test passes if a console warning is generated, noting the access violation.



--------
Frame: 'testframe'
--------
